// auth-service/authentication/src/models/token.interface.d.ts
export interface IJWTTokenInfoPublic extends IJWTTokenClaimsAndPayloadPublic {
    description: string;
}


/**
 * JWT Token Claim Values that need to be set
 * (meaning not exposed to the front-end and will
 *  be automatically set by back-end)
 *
 * @export
 * @interface IJWTTokenClaims
 */
export interface IJWTTokenClaimsAndPayloadPublic extends IJWTTokenPayloadPublic, IJWTTokenClaimsPublic { }


/**
 * JWT Token Claims and Payload that includes fields set by the back-end
 *
 * @export
 * @interface IJWTTokenClaimsAndPayloadSecure
 * @extends {IJWTTokenPayloadSecure}
 * @extends {IJWTTokenClaimsSecure}
 */
export interface IJWTTokenClaimsAndPayloadSecure extends IJWTTokenPayloadSecure, IJWTTokenClaimsSecure { }

export interface IJWTTokenClaimsPublic {

    // decrypted jwt token id
    // (firebase id to prevent collision)
    jti: string;

    // aud claim will be used to store **Participant's ID in the Participant Registry**
    // NOTE 1: Participant Id per PR (different than firebase participant id)
    // NOTE 2: was originally using AUD for endpoints but this
    // does not allow for string[] of endpoints
    aud: string; // JSON.stringify

}

/**
 * These are secure (meaning not exposed to the front-end)
 *
 * @export
 * @interface IJWTTokenClaimsSecure
 */
export interface IJWTTokenClaimsSecure extends IJWTTokenClaimsPublic {

    // to be set by back-end
    // issued at
    iat?: number;

    // expired time
    exp?: number;
    nbf?: number;

    // sub will be used to store the institution id per firebase user
    // Note:  **AUD** will be used to store the participant Id per the PR
    // this will be set via the header in the post request "x-iid"
    sub: string;

    // claim included in token header to lookup secret
    kid?: string;

    // not being used
    // iss?: string;
}

/**
 * JWT Token Payload that needs to be set
 * characters kept to a minimum for faster decryption
 *
 * @export
 * @interface IJWTTokenPayload
 */
export interface IJWTTokenPayloadPublic {

    // allowable 'human readable' stellar accounts
    acc: string[]; // JSON.stringify

    // versions
    ver: string;

    // allowable ips
    ips: string[];

    // environment this token can be used for.
    // generally pulled from environment variable
    env: string;

    // endpoints: use enp in middleware to check endpoints
    // instead aud claim so that access info is consistent
    // with where other access rights info is stored in payload,
    // AUD would require JSON.stringify to convert to a string
    // since audience is defined in spec as a string and NOT string[]
    enp: string[];

}

/**
 * This is secure (meaning not exposed to the front-end in jwt_info)
 *
 * @export
 * @interface IJWTTokenPayloadSecure
 * @extends {IJWTTokenPayloadPublic}
 */
export interface IJWTTokenPayloadSecure extends IJWTTokenPayloadPublic {
    // increment count watermark
    // increments by 1 each time the token is refreshed
    // if 0 then the token has never been initialized
    // maximum expiration time on a un-initialize token should be
    // less than 24hrs for increased security purposes, then the token should
    // be refreshed every 15 minutes thereafter
    n: number
}